requirements.txt
aiohttp==3.10.5
fastapi==0.114.1
notion_client==2.2.1
protobuf==5.28.1
pydantic==2.9.1
python-dotenv==1.0.1
Requests==2.32.3
